Convert all boot files to TS.
