+++ client/src/boot/apollo/dataIdFromObject.ts
@@ -1,4 +1,6 @@
-const dataIdFromObject = (o) => {
+import {DataObject} from "../../types/DataObject";
+
+const dataIdFromObject = (o: DataObject): string|null => {
   const dataId = o.id || o.ID;
   if (dataId && dataId >= 0 && o.__typename) {
     return `${o.__typename}:${dataId}`;
