+++ client/src/types/DataObject.ts
@@ -1,0 +1,5 @@
+export interface DataObject {
+    ID?: number|string,
+    id?: number|string,
+    __typename?: string,
+}
