+++ package.json
@@ -99,6 +99,7 @@
     "@storybook/addon-options": "^3.4.10",
     "@storybook/addons": "^3.4.10",
     "@storybook/react": "^3.4.10",
+    "@types/isomorphic-fetch": "^0.0.35",
     "@types/jest": "^24.0.5",
     "@types/jquery": "^3.3.29",
     "@typescript-eslint/eslint-plugin": "^1.3.0",
